requirements.txt
# Core dependencies
pandas>=1.5.0
requests>=2.28.0
openpyxl>=3.0.10
pyyaml>=6.0

# String matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Data analysis and visualization
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Progress bars and utilities
tqdm>=4.64.0
colorama>=0.4.5
